+++ .eslintrc.js
@@ -1,0 +1,13 @@
+module.exports = {
+  env: {
+    es6: true,
+    mocha: true
+  },
+  // "plugins": [
+  //   "security"
+  // ],
+  extends: [
+    'standard'
+    // "plugin:security/recommended"
+  ]
+};
