+++ LICENSE
@@ -1,0 +1,1 @@
+(c) Copyright 2020 Zynch, todos los derechos reservados.
