+++ index.js
@@ -1,0 +1,8 @@
+'use strict';
+
+const server = require('./server');
+
+/**
+ * Start server
+ */
+server.listen();
