+++ package.json
@@ -1,26 +1,40 @@
 {
-    "name": "07-restserver",
-    "version": "1.0.0",
-    "description": "",
-    "main": "index.js",
-    "scripts": {
-        "start": "node server/server.js",
-        "test": "echo \"Error: no test specified\" && exit 1"
-    },
-    "author": "",
-    "license": "ISC",
-    "dependencies": {
-        "axios": "^0.19.1",
-        "bcrypt": "^3.0.7",
-        "body-parser": "^1.19.0",
-        "express": "^4.17.1",
-        "hbs": "^4.1.0",
-        "https": "^1.0.0",
-        "jsonwebtoken": "^8.5.1",
-        "mongoose": "^5.8.7",
-        "mongoose-unique-validator": "^2.0.3",
-        "nodemailer": "^6.4.2",
-        "request": "^2.88.0",
-        "underscore": "^1.9.2"
-    }
+  "name": "zynch-middleware",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "start": "export NODE=ENV=development && npm run dev",
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "dev": "export NODE_ENV=development && nodemon index.js",
+    "production": "export NODE_ENV=production && nodemon index.js"
+  },
+  "author": "SpaceBar",
+  "license": "SEE LICENSE IN LICENSE",
+  "private": true,
+  "dependencies": {
+    "axios": "^0.19.1",
+    "bcrypt": "^3.0.7",
+    "body-parser": "^1.19.0",
+    "cors": "^2.8.5",
+    "dotenv": "^8.2.0",
+    "express": "^4.17.1",
+    "express-rate-limit": "^5.0.0",
+    "hbs": "^4.1.0",
+    "helmet": "^3.21.2",
+    "http": "0.0.0",
+    "https": "^1.0.0",
+    "jsonwebtoken": "^8.5.1",
+    "mongoose": "^5.8.7",
+    "mongoose-unique-validator": "^2.0.3",
+    "nodemailer": "^6.4.2",
+    "request": "^2.88.0",
+    "underscore": "^1.9.2",
+    "winston": "^3.2.1"
+  },
+  "devDependencies": {
+    "chai": "^4.2.0",
+    "mocha": "^7.0.0",
+    "nodemon": "^2.0.2"
+  }
 }
