+++ server/index.js
@@ -1,0 +1,8 @@
+'use strict';
+
+const { listen, close } = require('./server');
+
+module.exports = {
+  listen,
+  close
+};
